вынес форму поиска в отдельный компонент
