+++ src/app/searchForm/searchForm.module.css
@@ -1,0 +1,3 @@
+#search-form{
+  display: flex;
+}
