+++ src/app/taskList/TaskList.module.css
@@ -13,7 +13,3 @@
   font-weight: bold;
   font-size: 1.5rem;
 }
-
-#search-form{
-  display: flex;
-}
